third_party/blink/web_tests/external/wpt/pointerevents/bugs/events_after_lostpointercapture_remove_setcapture_node-expected.txt

This is a testharness.js-based test.
[FAIL] setPointerCapture target removed by lostpointercapture
  assert_equals: expected "lostpointercapture@target0, pointerup@button, pointerup@button" but got "lostpointercapture@target0, lostpointercapture@target0, pointerup@target1, pointerup@button"
Harness: the test ran to completion.
